refactor: Activity를 Fragment로 전환
AllStampFragment에서 화면을 보여줘야하는 이벤트를 한번 더 해야하기에 리소스를 잡아먹는 Activity대신 Fragment로 전환

issue: #208
