+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/stamp/StampActivity.kt
@@ -2,29 +2,18 @@ package com.nbcfinalteam2.ddaraogae.presentation.ui.stamp
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.updatePadding
-import androidx.lifecycle.flowWithLifecycle
-import androidx.lifecycle.lifecycleScope
+import com.nbcfinalteam2.ddaraogae.R
 import com.nbcfinalteam2.ddaraogae.databinding.ActivityStampBinding
-import com.nbcfinalteam2.ddaraogae.domain.bus.ItemChangedEventBus
-import com.nbcfinalteam2.ddaraogae.presentation.model.DefaultEvent
-import com.nbcfinalteam2.ddaraogae.presentation.util.ToastMaker
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.launch
-import javax.inject.Inject
 
 @AndroidEntryPoint
 class StampActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityStampBinding
-    private val allStampViewModel: AllStampViewModel by viewModels()
-    private lateinit var allStampAdapter: AllStampAdapter
-    @Inject lateinit var itemChangedEventBus: ItemChangedEventBus
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -32,9 +21,11 @@ class StampActivity : AppCompatActivity() {
         binding = ActivityStampBinding.inflate(layoutInflater)
         setContentView(binding.root)
         uiSetting()
-        setupAdapter()
-        setupViewModel()
-        setupListener()
+        if (savedInstanceState == null) {
+            supportFragmentManager.beginTransaction()
+                .replace(R.id.fragment_container, AllStampFragment())
+                .commitNow()
+        }
     }
 
     private fun uiSetting() {
@@ -44,39 +35,6 @@ class StampActivity : AppCompatActivity() {
             WindowInsetsCompat.CONSUMED
         }
     }
-
-    private fun setupAdapter() {
-        allStampAdapter = AllStampAdapter()
-        binding.rvStampArea.adapter = allStampAdapter
-    }
-
-    private fun setupViewModel() {
-        lifecycleScope.launch {
-            allStampViewModel.stampListState.collectLatest { stampList ->
-                allStampAdapter.submitList(stampList)
-            }
-        }
-
-        lifecycleScope.launch {
-            allStampViewModel.loadStampEvent.flowWithLifecycle(lifecycle).collectLatest { event ->
-                when (event) {
-                    is DefaultEvent.Failure -> ToastMaker.make(this@StampActivity, event.msg)
-                    DefaultEvent.Success -> {}
-                }
-            }
-        }
-
-        lifecycleScope.launch{
-            itemChangedEventBus.itemChangedEvent.flowWithLifecycle(lifecycle).collectLatest {
-                allStampViewModel.loadStampList()
-            }
-        }
-    }
-
-    private fun setupListener() {
-        binding.btnBack.setOnClickListener {
-            finish()
-        }
-    }
 }
+
